🔨Refactor: 코드 리팩토링

+++ src/main/kotlin/com/teamsparta/buysell/domain/member/controller/AccountController.kt
@@ -29,7 +29,7 @@ class AccountController(
 
     @PatchMapping
     fun chargeAccount(
-        @RequestParam money:Int,
+        @RequestParam money:Long,
         @AuthenticationPrincipal userPrincipal: UserPrincipal
     ): ResponseEntity<AccountResponse> {
         return ResponseEntity
