+++ src/main/kotlin/com/teamsparta/buysell/domain/member/model/Account.kt
@@ -7,18 +7,18 @@ import jakarta.persistence.*
 class Account (
 
     @Column(name = "account_balance")
-    var accountBalance: Int = 0
+    var accountBalance: Long = 0
 
 ){
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     var id: Int? = null
 
-    fun depositToAccount(money: Int){
+    fun depositToAccount(money: Long){
         accountBalance += money
     }
 
-    fun payment(price: Int){
+    fun payment(price: Long){
         if(availableForPurchase(price)){ //결제 가능할 때
             accountBalance -= price
         } else{ //결제 불가능할 때 에러를 던진다
@@ -26,7 +26,7 @@ class Account (
         }
     }
 
-    fun availableForPurchase(price: Int): Boolean{
+    fun availableForPurchase(price: Long): Boolean{
         return accountBalance > price
     }
 }
